Move all registration stuff in theirs own "engine" package
